React color picker from react-color working as a custom component
