+++ src/js/components/widget-color-picker/sketch-fields.react.js
@@ -1,0 +1,71 @@
+
+import React from 'react'
+import ReactCSS, {Component} from 'reactcss'
+import { EditableInput } from 'react-color/lib/components/common'
+import color from 'react-color/lib/helpers/color'
+
+import shallowCompare from 'react-addons-shallow-compare'
+
+export default class SketchFields extends React.Component {
+    constructor (props) {
+        super(props);
+    }
+
+  shouldComponentUpdate () {
+    return shallowCompare.bind(this, this, arguments[0], arguments[1]);
+  }
+
+  handleChange (data) {
+    if (data.hex) {
+      color.isValidHex(data.hex) && this.props.onChange({
+        hex: data.hex,
+        source: 'hex',
+      })
+    } else if (data.r || data.g || data.b) {
+      this.props.onChange({
+        r: data.r || this.props.rgb.r,
+        g: data.g || this.props.rgb.g,
+        b: data.b || this.props.rgb.b,
+        a: this.props.rgb.a,
+        source: 'rgb',
+      })
+    } else if (data.a) {
+      if (data.a < 0) {
+        data.a = 0
+      } else if (data.a > 100) {
+        data.a = 100
+      }
+
+      data.a = data.a / 100
+      this.props.onChange({
+        h: this.props.hsl.h,
+        s: this.props.hsl.s,
+        l: this.props.hsl.l,
+        a: data.a,
+        source: 'rgb',
+      })
+    }
+  }
+
+  render() {
+    return (
+      <div className="fields flexbox-fix">
+        <div className="double">
+          <EditableInput className="input" label="hex" value={ this.props.hex.replace('#', '') } onChange={ this.handleChange }/>
+        </div>
+        <div className="single">
+          <EditableInput className="input" label="r" value={ this.props.rgb.r } onChange={ this.handleChange } dragLabel="true" dragMax="255"/>
+        </div>
+        <div className="single">
+          <EditableInput className="input" label="g" value={ this.props.rgb.g } onChange={ this.handleChange } dragLabel="true" dragMax="255"/>
+        </div>
+        <div className="single">
+          <EditableInput className="input" label="b" value={ this.props.rgb.b } onChange={ this.handleChange } dragLabel="true" dragMax="255"/>
+        </div>
+        <div className="alpha3">
+          <EditableInput is="input" label="a" value={ Math.round(this.props.rgb.a * 100) } onChange={ this.handleChange } dragLabel="true" dragMax="100"/>
+        </div>
+      </div>
+    )
+  }
+}
